requirements.txt
# Live Transcriber - Apple Silicon Optimized
# Real-time speaker diarization and keyword tracking

# Core UI
streamlit>=1.31.0

# Audio Processing
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.11.0

# MLX-Whisper for Apple Silicon optimized transcription
mlx>=0.5.0
mlx-whisper>=0.1.0

# Speaker Diarization
pyannote.audio>=3.1.0
torch>=2.1.0
torchaudio>=2.1.0

# Hugging Face for model access
huggingface-hub>=0.20.0

# Utilities
python-dotenv>=1.0.0
